Minor code changes
