+++ header/Info.hpp
@@ -1,0 +1,6 @@
+#ifndef INFO_HPP
+#define INFO_HPP
+
+void about();
+
+#endif
